Remove repeated line
I've removed the line: It’s important to be able to say what you plan to do, clearly and concisely, in a short form, to anyone, on a moment’s notice! t. - as it's verbatim stated in the paragraph under Your Vision Drives Your Project, with the t just hanging at the end of that line appears to just be a copy paste error.